Reorganize transformers package and fully test dataframes
- Expose all transformers from transformers/__init__.py
- Move StandardScalerWithDOF into _base.py module
- Add np.asarray() as necessary
