+++ src/sknnr/transformers/__init__.py
@@ -1,3 +1,5 @@
-from ._standard_scaler_dof import StandardScalerWithDOF
+from ._base import StandardScalerWithDOF
+from ._cca_transformer import CCATransformer
+from ._mahalanobis_transformer import MahalanobisTransformer
 
-__all__ = ["StandardScalerWithDOF"]
+__all__ = ["StandardScalerWithDOF", "CCATransformer", "MahalanobisTransformer"]
